Crate mentat_db

Re-exports

pub use errors::DbError;

pub use errors::DbErrorKind;

pub use errors::Result;

pub use errors::SchemaConstraintViolation;

pub use entids::DB_SCHEMA_CORE;

pub use db::TypedSQLValue;

pub use db::new_connection;

pub use tx_observer::InProgressObserverTransactWatcher;

pub use tx_observer::TxObservationService;

pub use tx_observer::TxObserver;

pub use types::AttributeSet;

pub use types::DB;

pub use types::PartitionMap;

pub use types::TransactableValue;

pub use types::TxReport;

Modules

cache

db

debug

entids

errors

internal_types
Types used only within the transactor. These should not be exposed outside of this crate.

tx_observer

types

Macros

bail

Structs

AttributeBuilder

Constants

CORE_SCHEMA_VERSION

TX0
The first transaction ID applied to the knowledge base.

USER0
This is the start of the :db.part/user partition.

Traits

AttributeValidation

TransactWatcher

Functions

repeat_values
Prepare an SQL VALUES block, like (?, ?, ?), (?, ?, ?).

to_namespaced_keyword

transact
Transact the given entities against the given SQLite conn, using the given metadata. If you want this work to occur inside a SQLite transaction, establish one on the connection prior to calling this function.

transact_terms
Just like transact, but accepts lower-level inputs to allow bypassing the parser interface.