Struct mentat_db::cache::AttributeCaches

pub struct AttributeCaches { /* fields omitted */ }

Methods

impl AttributeCaches

pub fn unregister_attribute<U>(&mut self, attribute: U) where
U: Into<Entid>,

impl AttributeCaches

pub fn forward_attribute_cache_for_attribute<'a, 's>(
&'a self,
schema: &'s Schema,
a: Entid
) -> Option<&'a AttributeCache>

Return a reference to the cache for the provided a, if a names an attribute that is cached in the forward direction. If a doesn't name an attribute, or it's not cached at all, or it's only cached in reverse (v to e, not e to v), None is returned.

pub fn extend_cache_for_entities_and_attributes<'s, 'c>(
&mut self,
schema: &'s Schema,
sqlite: &'c Connection,
attrs: AttributeSpec,
entities: &Vec<Entid>
) -> Result<()>

Fetch the requested entities and attributes from the store and put them in the cache. The caller is responsible for ensuring that entities is unique. Attributes for which every entity is already cached will not be processed again.

pub fn make_cache_for_entities_and_attributes<'s, 'c>(
schema: &'s Schema,
sqlite: &'c Connection,
attrs: AttributeSpec,
entities: &Vec<Entid>
) -> Result<AttributeCaches>

Fetch the requested entities and attributes and put them in a new cache. The caller is responsible for ensuring that entities is unique.