Module mentat_db::db

Enums

SearchType

Constants

CURRENT_VERSION
Version history:

Traits

MentatStoring
MentatStoring will be the trait that encapsulates the storage layer. It is consumed by the transaction processing layer.

PartitionMapping

TypedSQLValue

Functions

create_current_version

create_empty_current_version
Do just enough work that either create_current_version or sync can populate the DB.

ensure_current_version

new_connection

read_db
Read the materialized views from the given SQL store and return a Mentat DB for querying and applying transactions.

update_metadata
Update the metadata materialized views based on the given metadata report.

update_partition_map
Update the current partition map materialized view.

Type Definitions

ReducedEntity
Internal representation of an [e a v added] datom, ready to be transacted against the store.