Module mentat_db::entids

Structs

IDENTS_SQL_LIST
Attributes that are "ident related". These might change the "idents" materialized view.

METADATA_SQL_LIST
Attributes that are "metadata" related. These might change one of the materialized views.

SCHEMA_SQL_LIST
Attributes that are "schema related". These might change the "schema" materialized view.

Constants

DB_ADD

DB_ALTER_ATTRIBUTE

DB_CARDINALITY

DB_CARDINALITY_MANY

DB_CARDINALITY_ONE

DB_DOC

DB_EXCISE

DB_EXCISE_ATTRS

DB_EXCISE_BEFORE

DB_EXCISE_BEFORE_T

DB_FULLTEXT

DB_IDENT

DB_INDEX

DB_INSTALL_ATTRIBUTE

DB_INSTALL_PARTITION

DB_INSTALL_VALUE_TYPE

DB_IS_COMPONENT

DB_NO_HISTORY

DB_PART_DB

DB_PART_TX

DB_PART_USER

DB_RETRACT

DB_SCHEMA_ATTRIBUTE

DB_SCHEMA_CORE

DB_SCHEMA_VERSION

DB_TX_INSTANT

DB_TYPE_BOOLEAN

DB_TYPE_BYTES

DB_TYPE_DOUBLE

DB_TYPE_INSTANT

DB_TYPE_KEYWORD

DB_TYPE_LONG

DB_TYPE_REF

DB_TYPE_STRING

DB_TYPE_URI

DB_TYPE_UUID

DB_UNIQUE

DB_UNIQUE_IDENTITY

DB_UNIQUE_VALUE

DB_VALUE_TYPE

Functions

is_a_schema_attribute
Return 'false' if the given attribute might be used to describe a schema attribute.

might_update_metadata
Return false if the given attribute will not change the metadata: recognized idents, schema, partitions in the partition map.