Enum mentat_db::errors::ErrorKind

pub enum ErrorKind {
    Msg(String),
    Rusqlite(Error),
    NotYetImplemented(String),
    BadValuePair(String, ValueType),
    BadSQLValuePair(Value, i32),
    BadBootstrapDefinition(String),
    BadSchemaAssertion(String),
    UnrecognizedIdent(String),
    UnrecognizedEntid(Entid),
    UnknownAttribute(Entid),
    CannotCacheNonUniqueAttributeInReverse(Entid),
    SchemaAlterationFailed(String),
    SchemaConstraintViolation(SchemaConstraintViolation),
    InputError(InputError),
    // some variants omitted
}

The kind of an error.

Variants
Msg(String)

A convenient variant for String.

Rusqlite(Error)NotYetImplemented(String)

We're just not done yet. Message that the feature is recognized but not yet implemented.

BadValuePair(String, ValueType)

We've been given a value that isn't the correct Mentat type.

BadSQLValuePair(Value, i32)

We've got corrupt data in the SQL store: a value and value_type_tag don't line up.

BadBootstrapDefinition(String)

A bootstrap definition couldn't be parsed or installed. This is a programmer error, not a runtime error.

BadSchemaAssertion(String)

A schema assertion couldn't be parsed.

UnrecognizedIdent(String)

An ident->entid mapping failed.

UnrecognizedEntid(Entid)

An entid->ident mapping failed. We also use this error if you try to transact an entid that we didn't allocate, in part because we blow the stack in error_chain if we define a new enum!

UnknownAttribute(Entid)CannotCacheNonUniqueAttributeInReverse(Entid)SchemaAlterationFailed(String)SchemaConstraintViolation(SchemaConstraintViolation)

A transaction tried to violate a constraint of the schema of the Mentat store.

InputError(InputError)

The transaction was malformed in some way (that was not recognized at parse time; for example, in a way that is schema-dependent).

Methods

impl ErrorKind

pub fn description(&self) -> &str

A string describing the error kind.