Function mentat_db::transact

pub fn transact<'conn, 'a, I, V, W>(
    conn: &'conn Connection, 
    partition_map: PartitionMap, 
    schema_for_mutation: &'a Schema, 
    schema: &'a Schema, 
    watcher: W, 
    entities: I
) -> Result<(TxReport, PartitionMap, Option<Schema>, W)> where
    I: IntoIterator<Item = Entity<V>>,
    V: TransactableValue,
    W: TransactWatcher, 

Transact the given entities against the given SQLite conn, using the given metadata. If you want this work to occur inside a SQLite transaction, establish one on the connection prior to calling this function.

This approach is explained in https://github.com/mozilla/mentat/wiki/Transacting.