Module mentat_db::internal_types

Types used only within the transactor. These should not be exposed outside of this crate.

Enums

LookupRefOrTempId
Internal representation of an entid on its way to resolution. We either have the simple case (a numeric entid), a lookup-ref that still needs to be resolved (an atomized [a v] pair), or a temp ID that needs to be upserted or allocated (an atomized tempid).

Term

Functions

replace_lookup_ref
Given a KnownEntidOr or a TypedValueOr, replace any internal LookupRef with the entid from the given map. Fail if any LookupRef cannot be replaced.

Type Definitions

KnownEntidOr

LookupRef

Population

TempIdHandle

TempIdMap

TermWithTempIds

TermWithTempIdsAndLookupRefs

TermWithoutTempIds

TypedValueOr