Module mentat_db::types

Structs

DB
Represents the metadata required to query from, or apply transactions to, a Mentat store.

DateTime
ISO 8601 combined date and time with time zone.

Partition
Represents one partition of the entid space.

PartitionMap
Map partition names to Partition instances.

Schema
Represents a Mentat schema.

Utc
The UTC time zone. This is the most efficient time zone when you don't need the local time. It is also used as an offset (which is also a dummy type).

Traits

TransactableValue
The transactor is tied to edn::ValueAndSpan right now, but in the future we'd like to support TypedValue directly for programmatic use. TransactableValue encapsulates the interface value types (i.e., values in the value place) need to support to be transacted.

Type Definitions

AVMap
Map [a v] pairs to existing entids.

AVPair
A pair [a v] in the store.

AttributeSet